requirements.txt
# 基础依赖
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.41.0

# 数据处理
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
jieba>=0.42.1

# 可视化和监控
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wandb>=0.15.0
tensorboard>=2.13.0

# Web界面
gradio>=3.40.0
streamlit>=1.25.0

# 评估指标
nltk>=3.8.0
rouge-score>=0.1.2
sacrebleu>=2.3.0

# 工具类
tqdm>=4.65.0
requests>=2.31.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# 可选：量化和优化
optimum>=1.12.0
auto-gptq>=0.4.0

# 开发工具
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0